feat: Removed google oauth, use ADC, and secrets.toml (#50)
- Removed google oauth to simplify the setup of the quickstart app
- Use gcloud ADC for google auth
- Updated readme.
- Switched from .env to streamlit's recommended secrets.toml file. 
- Setup CI to handle gcloud ADC and secrets.toml file. 
- The CI secrets.toml file generation can be leaked if an external
contributor runs the workflows, however 1) the file does not contain
actual secrets, only cloud project id. It will not contain any looker
secrets
- External contributors cannot run workflows without member/owner
approval

Multiple users had issues with setting up google Oauth. Even though the
quickstart used google oauth, it also used Looker client/secret, which
was inconsistent. We looked into implementing looker oauth in addition
to google oauth for consistency, but that would have required
implementing cookie management and maintaining 2 Oauth flows in one app.
In the end, we decided the oauth flow(s) complexity is not worth it,
given we want the users to be able to 1) use the quickstart as
easily/fast as possible 2) understand the app logic easily.

Closes:
https://github.com/looker-open-source/ca-api-quickstarts/issues/34